Minor update to object.h, object.c and klass.h.

+++ src/vm/object.h
@@ -239,7 +239,7 @@ struct ObjClass {
     struct ObjClass* superclass;
     ValueArray traits;
     bool isNative;
-    uint8_t interceptors;
+    uint16_t interceptors;
     IDMap indexes;
     ValueArray fields;
     Table methods;
